+++ crates/aiken-project/src/blueprint/mod.rs
@@ -1,8 +1,11 @@
 pub mod definitions;
 pub mod error;
+mod memo_program;
 pub mod parameter;
 pub mod schema;
 pub mod validator;
+
+pub use error::Error;
 
 use crate::{
     config::{self, Config},
@@ -10,7 +13,6 @@ use crate::{
 };
 use aiken_lang::gen_uplc::CodeGenerator;
 use definitions::Definitions;
-use error::Error;
 use schema::{Annotated, Schema};
 use std::fmt::Debug;
 use validator::Validator;