+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## v1.0.27-alpha - unreleased
+
+### Added
+
+- **aiken-lsp**: hover and goto definition support on list tail. @rvcas
+- **aiken-lsp**: hover on prop test via expression. @rvcas
+- **aiken**: added export command that exporting of regular function definitons. @rvcas
+- **aiken-lang**: a new way to emit logs that don't get erased. @micahkendall
+
+### Fixed
+
+- **aiken-lang**: formatter should not erase `pub` on validators. @rvcas
+- **aiken-lang**: error on using tuple index when a tuple is returned by a generic function. @rvcas
+
 ## v1.0.26-alpha - 2024-03-25
 
 ### Fixed